SwiftPM 6.4 support - refactor existing methods
